requirements.txt
Django==3.2.9
django-extensions==3.1.5
djangorestframework==3.12.4
PyJWT==2.3.0
six==1.16.0
django-cors-headers==3.10.1
